+++ src/evo/deterministicmns.cpp
@@ -6,8 +6,11 @@
 #include <evo/dmnstate.h>
 #include <evo/specialtx.h>
 #include <evo/simplifiedmns.h>
+#include <evo/evodb.h>
 #include <llmq/commitment.h>
 #include <llmq/utils.h>
+#include <llmq/complex_utils.h>
+#include <llmq/params.h>
 #include <evo/providertx.h>
 
 #include <base58.h>
@@ -876,7 +879,7 @@ void CDeterministicMNManager::HandleQuorumCommitment(const llmq::CFinalCommitmen
 {
     // The commitment has already been validated at this point, so it's safe to use members of it
 
-    auto members = llmq::CLLMQUtils::GetAllQuorumMembers(qc.llmqType, pQuorumBaseBlockIndex);
+    auto members = llmq::CLLMQComplexUtils::GetAllQuorumMembers(qc.llmqType, pQuorumBaseBlockIndex);
 
     for (size_t i = 0; i < members.size(); i++) {
         if (!mnList.HasMN(members[i]->proTxHash)) {
