+++ src/llmq/debug.cpp
@@ -10,6 +10,7 @@
 
 #include <evo/deterministicmns.h>
 #include <llmq/utils.h>
+#include <llmq/complex_utils.h>
 #include <util/irange.h>
 
 namespace llmq
@@ -28,7 +29,7 @@ UniValue CDKGDebugSessionStatus::ToJson(int quorumIndex, int detailLevel) const
     if (detailLevel == 2) {
         const CBlockIndex* pindex = WITH_LOCK(cs_main, return LookupBlockIndex(quorumHash));
         if (pindex != nullptr) {
-            dmnMembers = CLLMQUtils::GetAllQuorumMembers(llmqType, pindex);
+            dmnMembers = CLLMQComplexUtils::GetAllQuorumMembers(llmqType, pindex);
         }
     }
