+++ src/util/hexstr.h
@@ -1,0 +1,13 @@
+// Copyright (c) 2022 The Dash Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_UTIL_GETUNIQUEPATH_H
+#define BITCOIN_UTIL_GETUNIQUEPATH_H
+
+#include <vector>
+#include <string>
+
+std::string BitsToHexStr(const std::vector<bool>& vBits);
+
+#endif // BITCOIN_UTIL_GETUNIQUEPATH_H
